pyproject.toml
[tool.poetry]
name = "gemini-rag-agent"
version = "0.1.0"
description = "Retrieval-Augmented Generation agent powered by Google Gemini"
authors = ["Zhang TongYan <example@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langchain-community = "^0.1.0"
langchain-google-genai = "^0.1.0"
google-generativeai = "^0.3.0"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.3"
flake8 = "^7.0"
isort = "^5.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
